Fix getWilayas by a straightforward read
